+++ client/web/src/enterprise/batches/batch-spec/edit/EditBatchSpecPage.tsx
@@ -7,7 +7,7 @@ import { useQuery } from '@sourcegraph/http-client'
 import { Settings, SettingsCascadeProps } from '@sourcegraph/shared/src/settings/settings'
 // import { useTemporarySetting } from '@sourcegraph/shared/src/settings/temporary/useTemporarySetting'
 import { ThemeProps } from '@sourcegraph/shared/src/theme'
-import { Button, Icon, LoadingSpinner, H4, Alert } from '@sourcegraph/wildcard'
+import { Button, Icon, LoadingSpinner, H3, H4, Alert } from '@sourcegraph/wildcard'
 
 import { HeroPage } from '../../../../components/HeroPage'
 import { useFeatureFlag } from '../../../../featureFlags/useFeatureFlag'
@@ -274,7 +274,9 @@ const MemoizedEditBatchSpecPageContent: React.FunctionComponent<
                 <div className={styles.form}>
                     <LibraryPane name={batchChange.name} onReplaceItem={editor.handleCodeChange} />
                     <div className={styles.editorContainer}>
-                        <H4 className={styles.header}>Batch spec</H4>
+                        <H4 as={H3} className={styles.header}>
+                            Batch spec
+                        </H4>
                         {executionAlert}
                         <MonacoBatchSpecEditor
                             autoFocus={true}