+++ client/web/src/enterprise/insights/core/backend/gql-backend/deserialization/create-insight-view.ts
@@ -6,11 +6,12 @@ import { SeriesSortDirection } from '@sourcegraph/shared/src/schema'
 import {
     InsightViewNode,
     SeriesSortMode,
+    GroupByField,
     TimeIntervalStepInput,
     TimeIntervalStepUnit,
 } from '../../../../../../graphql-operations'
 import { parseSeriesDisplayOptions } from '../../../../components/insights-view-grid/components/backend-insight/components/drill-down-filters-panel/drill-down-filters/utils'
-import { Insight, InsightExecutionType, InsightType } from '../../../types'
+import { ComputeInsight, Insight, InsightExecutionType, InsightType } from '../../../types'
 import { BaseInsight } from '../../../types/insight/common'
 import { MAX_NUMBER_OF_SERIES } from '../methods/get-backend-insight-data/deserializators'
 
@@ -21,8 +22,9 @@ import { MAX_NUMBER_OF_SERIES } from '../methods/get-backend-insight-data/deseri
  * api for insights.
  */
 export const createInsightView = (insight: InsightViewNode): Insight => {
-    const baseInsight: Omit<BaseInsight, 'title' | 'type' | 'executionType'> = {
+    const baseInsight: Omit<BaseInsight, 'type' | 'executionType'> = {
         id: insight.id,
+        title: insight.presentation.title,
         isFrozen: insight.isFrozen,
         dashboardReferenceCount: insight.dashboardReferenceCount,
         seriesDisplayOptions: parseSeriesDisplayOptions(insight.seriesCount, insight.appliedSeriesDisplayOptions),
@@ -34,9 +36,12 @@ export const createInsightView = (insight: InsightViewNode): Insight => {
 
     switch (insight.presentation.__typename) {
         case 'LineChartInsightViewPresentation': {
+            const isComputeInsight = insight.dataSeriesDefinitions.some(series => series.groupBy)
             const isCaptureGroupInsight = insight.dataSeriesDefinitions.some(
                 series => series.generatedFromCaptureGroups
             )
+
+            const { appliedFilters } = insight
             // We do not support different time scope for different series at the moment
             const step = getDurationFromStep(insight.dataSeriesDefinitions[0].timeScope)
             const repositories = uniq(
@@ -59,13 +64,12 @@ export const createInsightView = (insight: InsightViewNode): Insight => {
             if (isCaptureGroupInsight) {
                 // It's safe because capture group insight always has only 1 data series
                 const { query } = insight.dataSeriesDefinitions[0] ?? {}
-                const { presentation, appliedFilters } = insight
+                const { appliedFilters } = insight
 
                 return {
                     ...baseInsight,
                     executionType: InsightExecutionType.Backend,
                     type: InsightType.CaptureGroup,
-                    title: presentation.title,
                     repositories,
                     query,
                     step,
@@ -94,13 +98,29 @@ export const createInsightView = (insight: InsightViewNode): Insight => {
                         : '',
             }))
 
-            const { presentation, appliedFilters } = insight
+            if (isComputeInsight) {
+                // It's safe because capture group insight always has only 1 data series
+                const { groupBy } = insight.dataSeriesDefinitions[0] ?? {}
+
+                return {
+                    ...baseInsight,
+                    executionType: InsightExecutionType.Backend,
+                    type: InsightType.Compute,
+                    groupBy: groupBy ?? GroupByField.REPO,
+                    repositories,
+                    series,
+                    filters: {
+                        includeRepoRegexp: appliedFilters.includeRepoRegex ?? '',
+                        excludeRepoRegexp: appliedFilters.excludeRepoRegex ?? '',
+                        context: appliedFilters.searchContexts?.[0] ?? '',
+                    },
+                } as ComputeInsight
+            }
 
             return {
                 ...baseInsight,
                 executionType: InsightExecutionType.Backend,
                 type: InsightType.SearchBased,
-                title: presentation.title,
                 repositories,
                 series,
                 step,