+++ client/web/src/integration/insights/fixtures/insights-metadata.ts
@@ -62,6 +62,7 @@ export const createJITMigrationToGQLInsightMetadataFixture = (options: InsightOp
                 unit: TimeIntervalStepUnit.WEEK,
                 value: 6,
             },
+            groupBy: null,
         },
         {
             __typename: 'SearchInsightDataSeriesDefinition',
@@ -78,6 +79,7 @@ export const createJITMigrationToGQLInsightMetadataFixture = (options: InsightOp
                 unit: TimeIntervalStepUnit.WEEK,
                 value: 6,
             },
+            groupBy: null,
         },
     ],
 })
@@ -125,6 +127,7 @@ export const STORYBOOK_GROWTH_INSIGHT_METADATA_FIXTURE: InsightViewNode = {
                 unit: TimeIntervalStepUnit.WEEK,
                 value: 6,
             },
+            groupBy: null,
         },
     ],
 }
@@ -164,6 +167,7 @@ export const SOURCEGRAPH_LANG_STATS_INSIGHT_METADATA_FIXTURE: InsightViewNode = 
             },
             isCalculated: false,
             generatedFromCaptureGroups: false,
+            groupBy: null,
             __typename: 'SearchInsightDataSeriesDefinition',
         },
     ],