+++ dev/sg/internal/check/runner.go
@@ -5,9 +5,8 @@ import (
 	"fmt"
 	"io"
 	"strings"
+	"sync"
 	"time"
-
-	"go.uber.org/atomic"
 
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/analytics"
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/std"
@@ -15,6 +14,8 @@ import (
 	"github.com/sourcegraph/sourcegraph/lib/group"
 	"github.com/sourcegraph/sourcegraph/lib/output"
 )
+
+type SuggestFunc[Args any] func(category string, c *Check[Args], err error) string
 
 type Runner[Args any] struct {
 	Input      io.Reader
@@ -34,6 +35,9 @@ type Runner[Args any] struct {
 	// Concurrency controls the maximum number of checks across categories to evaluate at
 	// the same time - defaults to 10.
 	Concurrency int
+	// SuggestOnCheckFailure can be implemented to prompt the user to try certain things
+	// if a check fails. The suggestion string can be in Markdown.
+	SuggestOnCheckFailure SuggestFunc[Args]
 }
 
 // NewRunner creates a Runner for executing checks and applying fixes in a variety of ways.
@@ -185,12 +189,34 @@ func (r *Runner[Args]) runAllCategoryChecks(ctx context.Context, args Args) *run
 
 		// used for progress bar - needs to be thread-safe since it can be updated from
 		// multiple categories at once.
-		checksDone           atomic.Float64
+		progressMu           sync.Mutex
+		checksDone           float64
 		updateChecksProgress = func() {
-			progress.SetValue(0, checksDone.Load()+1)
-			checksDone.Add(1)
+			progressMu.Lock()
+			defer progressMu.Unlock()
+
+			checksDone += 1
+			progress.SetValue(0, checksDone)
+		}
+		updateCheckSkipped = func(i int, checkName string, err error) {
+			progressMu.Lock()
+			defer progressMu.Unlock()
+
+			progress.StatusBarUpdatef(i, "Check %s skipped: %s", checkName, err.Error())
+		}
+		updateCheckFailed = func(i int, checkName string, err error) {
+			progressMu.Lock()
+			defer progressMu.Unlock()
+
+			errParts := strings.SplitN(err.Error(), "\n", 2)
+			if len(errParts) > 2 {
+				// truncate to one line - writing multple lines causes some jank
+				errParts[0] += " ..."
+			}
+			progress.StatusBarFailf(i, "Check %s failed: %s", checkName, errParts[0])
 		}
 	)
+
 	for i, category := range r.Categories {
 		progress.StatusBarUpdatef(i, "Running checks...")
 
@@ -217,7 +243,8 @@ func (r *Runner[Args]) runAllCategoryChecks(ctx context.Context, args Args) *run
 				// run checks concurrently
 				checksGroup.Go(func() (event string, err error) {
 					if err := check.IsEnabled(ctx, args); err != nil {
-						progress.StatusBarUpdatef(i, "Check %s skipped: %s", check.Name, err.Error())
+						updateCheckSkipped(i, check.Name, err)
+
 						return "skipped", nil
 					}
 
@@ -225,12 +252,8 @@ func (r *Runner[Args]) runAllCategoryChecks(ctx context.Context, args Args) *run
 					// progress to discard.
 					var updateOutput strings.Builder
 					if err := check.Update(ctx, std.NewFixedOutput(&updateOutput, true), args); err != nil {
-						errParts := strings.SplitN(err.Error(), "\n", 2)
-						if len(errParts) > 2 {
-							// truncate to one line - writing multple lines causes some jank
-							errParts[0] += " ..."
-						}
-						progress.StatusBarFailf(i, "Check %s failed: %s", check.Name, errParts[0])
+						updateCheckFailed(i, check.Name, err)
+
 						check.cachedCheckOutput = updateOutput.String()
 						return "error", err
 					}
@@ -300,12 +323,21 @@ func (r *Runner[Args]) runAllCategoryChecks(ctx context.Context, args Args) *run
 			for _, check := range category.Checks {
 				if check.cachedCheckErr != nil {
 					// Slightly different formatting for each destination
+					var suggestion string
+					if r.SuggestOnCheckFailure != nil {
+						suggestion = r.SuggestOnCheckFailure(category.Name, check, check.cachedCheckErr)
+					}
 
 					// Write the terminal summary to an indented block
 					var style = output.CombineStyles(output.StyleBold, output.StyleFailure)
 					block := r.Output.Block(output.Linef(output.EmojiFailure, style, check.Name))
 					block.Writef("%s\n", check.cachedCheckErr)
-					block.Writef("%s\n", check.cachedCheckOutput)
+					if check.cachedCheckOutput != "" {
+						block.Writef("%s\n", check.cachedCheckOutput)
+					}
+					if suggestion != "" {
+						block.WriteLine(output.Styled(output.StyleSuggestion, suggestion))
+					}
 					block.Close()
 
 					// Build the markdown for the annotation summary
@@ -317,6 +349,10 @@ func (r *Runner[Args]) runAllCategoryChecks(ctx context.Context, args Args) *run
 							strings.TrimSpace(check.cachedCheckOutput))
 
 						annotationSummary += outputMarkdown
+					}
+
+					if suggestion != "" {
+						annotationSummary += fmt.Sprintf("\n\n%s", suggestion)
 					}
 
 					if r.GenerateAnnotations {