+++ enterprise/cmd/worker/internal/permissions/bitbucket_projects.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
-	"net/url"
 	"sort"
 	"strconv"
 	"time"
@@ -14,8 +13,6 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/sourcegraph/log"
-	"github.com/sourcegraph/sourcegraph/internal/conf/reposource"
-
 	"github.com/sourcegraph/sourcegraph/internal/errcode"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/globals"
@@ -113,13 +110,13 @@ func (h *bitbucketProjectPermissionsHandler) Handle(ctx context.Context, logger 
 	projectKey := workerJob.ProjectKey
 
 	// These repos are fetched from Bitbucket, therefore their IDs are Bitbucket IDs
-	// and we need to search for these repos in frontend DB using the name
+	// and we need to search for these repos in frontend DB to get Sourcegraph internal IDs
 	bitbucketRepos, err := client.ProjectRepos(ctx, projectKey)
 	if err != nil {
 		return errors.Wrapf(err, "failed to list repositories of Bitbucket Project %q", projectKey)
 	}
 
-	repoIDs, err := h.getRepoIDsByNames(ctx, svc, projectKey, bitbucketRepos)
+	repoIDs, err := h.getRepoIDsByNames(ctx, svc, bitbucketRepos)
 	if err != nil {
 		return errors.Wrap(err, "failed to get gitserver repos from the database")
 	}
@@ -182,51 +179,37 @@ func (h *bitbucketProjectPermissionsHandler) setReposUnrestricted(ctx context.Co
 	return nil
 }
 
-// getRepoIDsByNames queries repo IDs from frontend database using repo names fetched from
+// getRepoIDsByNames queries repo IDs from frontend database using external repo IDs fetched from
 // Bitbucket code host.
-func (h *bitbucketProjectPermissionsHandler) getRepoIDsByNames(ctx context.Context, svc *types.ExternalService, projectKey string, repos []*bitbucketserver.Repo) ([]api.RepoID, error) {
+func (h *bitbucketProjectPermissionsHandler) getRepoIDsByNames(ctx context.Context, svc *types.ExternalService, repos []*bitbucketserver.Repo) ([]api.RepoID, error) {
 	count := len(repos)
 	IDs := make([]api.RepoID, 0, count)
 	if count == 0 {
 		return IDs, nil
 	}
 
-	// unmarshalling external service config
-	var cfg schema.BitbucketServerConnection
-	if err := jsonc.Unmarshal(svc.Config, &cfg); err != nil {
-		return nil, errors.Errorf("external service id=%d config error: %s", svc.ID, err)
-	}
-
-	// parsing the hostname from the URL
-	parsedURL, err := url.Parse(cfg.Url)
-	if err != nil {
-		return nil, errors.Errorf("error during parsing external service URL", err)
-	}
-	hostname := parsedURL.Hostname()
-
-	names := make([]api.RepoName, 0, count)
+	specs := make([]api.ExternalRepoSpec, 0, count)
+	extSvcType := extsvc.KindToType(svc.Kind)
+	extSvcID := strconv.FormatInt(svc.ID, 10)
 	for _, repo := range repos {
-		// this is how repo names are composed before creating repos in `repo` table
-		// we are reconstructing this name for successful pattern matching
-		name := reposource.BitbucketServerRepoName(
-			cfg.RepositoryPathPattern,
-			hostname,
-			projectKey,
-			repo.Slug,
-		)
+		// using external ID, external service type and external service ID of the repo to find it
+		spec := api.ExternalRepoSpec{
+			ID:          strconv.Itoa(repo.ID),
+			ServiceType: extSvcType,
+			ServiceID:   extSvcID,
+		}
 
-		names = append(names, name)
+		specs = append(specs, spec)
 	}
 
-	// searching for repos by names
-	gitserverRepos, err := h.db.GitserverRepos().GetByNames(ctx, names...)
+	foundRepos, err := h.db.Repos().List(ctx, database.ReposListOptions{ExternalRepos: specs})
 	if err != nil {
 		return nil, err
 	}
 
 	// mapping repos to repo IDs
-	for _, gitserverRepo := range gitserverRepos {
-		IDs = append(IDs, gitserverRepo.RepoID)
+	for _, foundRepo := range foundRepos {
+		IDs = append(IDs, foundRepo.ID)
 	}
 
 	return IDs, nil